Work on device system.

+++ modules/ahci/ahci.h
@@ -143,4 +143,4 @@ fis_recv_t;
 
 void ahci_setup(uptr abar);
 
-bool ahci_read(port_t* port, u64 lba, void *buf, u32 count);
+bool ahci_read(port_t* port, u64 lba, void *buf, u64 count);
